Rename graph to directed graph
