+++ src/CombatDicesTeam.Graphs/README.md
@@ -1,3 +1,3 @@
 # CombatDicesTeam.Graphs
 
-The simplest library to work with graph.
+The simplest library to work with graph. The library was made for the indie game devs, so as not to pull monstrous enterprise solutions for working with graphs into small pet-games.
